requirements.txt
streamlit
websocket-client
